+++ lib/src/code_generator/struc.dart
@@ -76,6 +76,13 @@ class Struc extends NoLookUpBinding {
     return array;
   }
 
+  String _getInlineArrayTypeString(Type type, Writer w) {
+    if (type.broadType == BroadType.ConstantArray) {
+      return '${w.ffiLibraryPrefix}.Array<${_getInlineArrayTypeString(type.child!, w)}>';
+    }
+    return type.getCType(w);
+  }
+
   List<Typedef>? _typedefDependencies;
   @override
   List<Typedef> getTypedefDependencies(Writer w) {
@@ -101,10 +108,6 @@ class Struc extends NoLookUpBinding {
       s.write(makeDartDoc(dartDoc!));
     }
 
-    final helpers = <ArrayHelper>[];
-
-    final expandedArrayItemPrefix = getUniqueExpandedArrayItemPrefix();
-
     /// Adding [enclosingClassName] because dart doesn't allow class member
     /// to have the same name as the class.
     final localUniqueNamer = UniqueNamer({enclosingClassName});
@@ -116,23 +119,15 @@ class Struc extends NoLookUpBinding {
     // Write class declaration.
     s.write(
         'class $enclosingClassName extends ${w.ffiLibraryPrefix}.${isOpaque ? 'Opaque' : 'Struct'}{\n');
+    const depth = '  ';
     for (final m in members) {
       final memberName = localUniqueNamer.makeUnique(m.name);
       if (m.type.broadType == BroadType.ConstantArray) {
-        // TODO(5): Remove array helpers when inline array support arives.
-        final arrayHelper = ArrayHelper(
-          helperClassGroupName:
-              '${w.arrayHelperClassPrefix}_${enclosingClassName}_$memberName',
-          elementType: m.type.getBaseArrayType(),
-          dimensions: _getArrayDimensionLengths(m.type),
-          name: memberName,
-          structName: enclosingClassName,
-          elementNamePrefix: '$expandedArrayItemPrefix${memberName}_item_',
-        );
-        s.write(arrayHelper.declarationString(w));
-        helpers.add(arrayHelper);
+        s.write(
+            '$depth@${w.ffiLibraryPrefix}.Array.multi(${_getArrayDimensionLengths(m.type)})\n');
+        s.write(
+            '${depth}external ${_getInlineArrayTypeString(m.type, w)} $memberName;\n\n');
       } else {
-        const depth = '  ';
         if (m.dartDoc != null) {
           s.write(depth + '/// ');
           s.writeAll(m.dartDoc!.split('\n'), '\n' + depth + '/// ');
@@ -146,27 +141,7 @@ class Struc extends NoLookUpBinding {
     }
     s.write('}\n\n');
 
-    for (final helper in helpers) {
-      s.write(helper.helperClassString(w));
-    }
-
     return BindingString(type: BindingStringType.struc, string: s.toString());
-  }
-
-  /// Gets a unique prefix in local namespace for expanded array items.
-  String getUniqueExpandedArrayItemPrefix() {
-    final base = '_unique';
-    var expandedArrayItemPrefix = base;
-    var suffixInt = 0;
-    for (var i = 0; i < members.length; i++) {
-      if (members[i].name.startsWith(expandedArrayItemPrefix)) {
-        // Not a unique prefix, start over with a new suffix.
-        i = -1;
-        suffixInt++;
-        expandedArrayItemPrefix = '$base$suffixInt';
-      }
-    }
-    return expandedArrayItemPrefix + '_';
   }
 }
 
@@ -182,138 +157,4 @@ class Member {
     required this.type,
     this.dartDoc,
   }) : originalName = originalName ?? name;
-}
-
-// Helper bindings for struct array.
-class ArrayHelper {
-  final Type elementType;
-  final List<int> dimensions;
-  final String? structName;
-
-  final String? name;
-  final String helperClassGroupName;
-  final String elementNamePrefix;
-
-  int? _expandedArrayLength;
-  int get expandedArrayLength {
-    if (_expandedArrayLength != null) return _expandedArrayLength!;
-
-    var arrayLength = 1;
-    for (final i in dimensions) {
-      arrayLength = arrayLength * i;
-    }
-    return arrayLength;
-  }
-
-  ArrayHelper({
-    required this.elementType,
-    required this.dimensions,
-    required this.structName,
-    required this.name,
-    required this.helperClassGroupName,
-    required this.elementNamePrefix,
-  });
-
-  /// Create declaration binding, added inside the struct binding.
-  String declarationString(Writer w) {
-    final s = StringBuffer();
-    final arrayDartType = elementType.getDartType(w);
-    final arrayCType = elementType.getCType(w);
-
-    for (var i = 0; i < expandedArrayLength; i++) {
-      if (elementType.isPrimitive) {
-        s.write('  @$arrayCType()\n');
-      }
-      s.write('  external $arrayDartType $elementNamePrefix$i;\n');
-    }
-
-    s.write('/// Helper for array `$name`.\n');
-    s.write(
-        '${helperClassGroupName}_level0 get $name => ${helperClassGroupName}_level0(this, $dimensions, 0, 0);\n');
-
-    return s.toString();
-  }
-
-  String helperClassString(Writer w) {
-    final s = StringBuffer();
-    final arrayType = elementType.getDartType(w);
-    for (var dim = 0; dim < dimensions.length; dim++) {
-      final helperClassName = '${helperClassGroupName}_level$dim';
-      final structIdentifier = '_struct';
-      final dimensionsIdentifier = 'dimensions';
-      final levelIdentifier = 'level';
-      final absoluteIndexIdentifier = '_absoluteIndex';
-      final checkBoundsFunctionIdentifier = '_checkBounds';
-      final legthIdentifier = 'length';
-
-      s.write('/// Helper for array `$name` in struct `$structName`.\n');
-
-      // Write class declaration.
-      s.write('class $helperClassName{\n');
-      s.write('final $structName $structIdentifier;\n');
-      s.write('final List<int> $dimensionsIdentifier;\n');
-      s.write('final int $levelIdentifier;\n');
-      s.write('final int $absoluteIndexIdentifier;\n');
-      s.write(
-          'int get $legthIdentifier => $dimensionsIdentifier[$levelIdentifier];\n');
-
-      // Write class constructor.
-      s.write(
-          '$helperClassName(this.$structIdentifier, this.$dimensionsIdentifier, this.$levelIdentifier, this.$absoluteIndexIdentifier);\n');
-
-      // Write checkBoundsFunction.
-      s.write('''
-  void $checkBoundsFunctionIdentifier(int index) {
-    if (index >= $legthIdentifier || index < 0) {
-      throw RangeError('Dimension \$$levelIdentifier: index not in range 0..\$$legthIdentifier exclusive.');
-    }
-  }
-  ''');
-      // If this isn't the last level.
-      if (dim + 1 != dimensions.length) {
-        // Override [] operator.
-        s.write('''
-  ${helperClassGroupName}_level${dim + 1} operator [](int index) {
-    $checkBoundsFunctionIdentifier(index);
-    var offset = index;
-    for (var i = level + 1; i < $dimensionsIdentifier.length; i++) {
-      offset *= $dimensionsIdentifier[i];
-    }
-    return ${helperClassGroupName}_level${dim + 1}(
-        $structIdentifier, $dimensionsIdentifier, $levelIdentifier + 1, $absoluteIndexIdentifier + offset);
-  }
-''');
-      } else {
-        // This is the last level, add switching logic here.
-        // Override [] operator.
-        s.write('$arrayType operator[](int index){\n');
-        s.write('$checkBoundsFunctionIdentifier(index);\n');
-        s.write('switch($absoluteIndexIdentifier+index){\n');
-        for (var i = 0; i < expandedArrayLength; i++) {
-          s.write('case $i:\n');
-          s.write('  return $structIdentifier.$elementNamePrefix$i;\n');
-        }
-        s.write('default:\n');
-        s.write("  throw Exception('Invalid Array Helper generated.');");
-        s.write('}\n');
-        s.write('}\n');
-
-        // Override []= operator.
-        s.write('void operator[]=(int index, $arrayType value){\n');
-        s.write('$checkBoundsFunctionIdentifier(index);\n');
-        s.write('switch($absoluteIndexIdentifier+index){\n');
-        for (var i = 0; i < expandedArrayLength; i++) {
-          s.write('case $i:\n');
-          s.write('  $structIdentifier.$elementNamePrefix$i = value;\n');
-          s.write('  break;\n');
-        }
-        s.write('default:\n');
-        s.write("  throw Exception('Invalid Array Helper generated.');\n");
-        s.write('}\n');
-        s.write('}\n');
-      }
-      s.write('}\n');
-    }
-    return s.toString();
-  }
 }