+++ test/native_test/native_test.dart
@@ -91,19 +91,20 @@ void main() {
     test('double', () {
       expect(bindings.Function1Double(0), 42.0);
     });
-    test('array-workaround: Order of access', () {
+    test('Array Test: Order of access', () {
       final struct1 = bindings.getStruct1();
       var expectedValue = 1;
-      for (var i = 0; i < struct1.ref.data.dimensions[0]; i++) {
-        for (var j = 0; j < struct1.ref.data.dimensions[1]; j++) {
-          for (var k = 0; k < struct1.ref.data.dimensions[2]; k++) {
+      final dimensions = [3, 1, 2];
+      for (var i = 0; i < dimensions[0]; i++) {
+        for (var j = 0; j < dimensions[1]; j++) {
+          for (var k = 0; k < dimensions[2]; k++) {
             expect(struct1.ref.data[i][j][k], expectedValue);
             expectedValue++;
           }
         }
       }
     });
-    test('array-workaround: Range Errors', () {
+    test('Array Workaround: Range Errors', () {
       final struct1 = bindings.getStruct1();
       // Index (get) above range.
       expect(