+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 3.0.0-beta.0
+- Added support for inline arrays in `Struct`s.
+- Remove config key `array-workaround`.
+- Remove deprecated key `llvm-lib` from config, Use `llvm-path` instead.
+
 # 2.5.0-beta.1
 - Added support for `Packed` structs. Packed annotations are generated
 automatically but can be overriden using `structs -> pack` config.